Introduce functional dependencies in Interpret to obviate inline type signatures.

+++ examples/Addition.hs
@@ -18,6 +18,6 @@ addition = do
 
 main = do
   let (worked, (a,b,c)) = evalNProgram RSat addition
-  putStrLn $ "a: " ++ show (a::Int)
-  putStrLn $ "b: " ++ show (b::Int)
-  putStrLn $ "c: " ++ show (c::Int)
+  putStrLn $ "a: " ++ show a
+  putStrLn $ "b: " ++ show b
+  putStrLn $ "c: " ++ show c
